fix: move popper content inside popper trigger
The Popper.content renders as a portal so this change will not interfare
with the UI. The content will just render in the document body.

This change was important because Primitives do not support multiple
children. Since Popper root needs to support asChild prop and root also
uses primitive div, this would mean adding multiple children, i.e.
Popper.Content and Popper.Trigger to Popper.Root would not render as
child. It would always render under a div wrapper.